[while] 100까지의 정수를 여러 개 입력 받아서 합, 평균, 입력 갯수 출력
0보다 작은 값이 입력되면 프로그램을 종료
100보다 큰 수는 합, 평균, 입력 갯수에서 제외

+++ Test_16.java
@@ -11,7 +11,7 @@ public class Test_16 {
 		int n;
 		int cnt = 0;
 		int sum = 0;
-		int average = 1;
+		double average = 1;
 		
 		while(true) {
 			System.out.print("정수 입력 : ");
